Enum trillium_api::Error

pub enum Error {
    ParseError {
        path: String,
        message: String,
    },
    IoError {
        kind: String,
        message: String,
    },
    UnsupportedMimeType {
        mime_type: String,
    },
    MissingContentType,
    Other {
        message: String,
    },
    FailureToNegotiateContent,
}

A serde-serializable error

Variants

ParseError

Fields

path: String

the path of the parse error, as provided by serde_path_to_error

message: String

the contents of the error

An error occurred in parsing the provided body content

IoError

Fields

kind: String

stringified std::io::ErrorKind

message: String

stringified std::io::Error

A transmission error occurred in the connection to the http client

UnsupportedMimeType

Fields

mime_type: String

the unsupported mime type

The client provided a content type that this library does not yet support

MissingContentType

The client did not provide a content-type

Other

Fields

message: String

A stringified error

Miscellaneous other errors โ please open an issue on trillium-api if you find yourself parsing the contents of this.

FailureToNegotiateContent

we were unable to find a content type that matches the Accept header. Please open an issue if youโd like an additional format to be supported