Struct trillium_api::ApiHandler

pub struct ApiHandler<F, OutputHandler, TryFromConn>(/* private fields */);

An interface layer built on trillium

This handler provides the capacity to extract various components of a conn such as deserializing a body, and supports returning handlers that will be called on the returned conn.

If ApiHandler encounters an error of any sort before the user-provided logic is executed, it will put an [Error] into the conn’s state. A default error handler is provided.

More documentation for this type is needed, hence the -rc semver on this crate

Implementations

impl<TryFromConnHandler, OutputHandler, Extracted> ApiHandler<TryFromConnHandler, OutputHandler, Extracted>
where TryFromConnHandler: for<'a> MutBorrowConn<'a, OutputHandler, Extracted>, OutputHandler: Handler, Extracted: TryFromConn,

pub fn new(api_handler: TryFromConnHandler) -> Self

constructs a new [ApiTryFromConnHandler] from the provided async fn(&mut conn, TryFromConn) -> impl Handler