Trait trillium_api::TryFromConn

pub trait TryFromConn: Send + Sync + Sized + 'static {
    type Error: Send + Sync + Sized + 'static;

    // Required method
    fn try_from_conn<'life0, 'async_trait>(
        conn: &'life0 mut Conn
    ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Like FromConn, but with an Error.

If you want to use this directly, Error needs to be Handler.

If Error is not Handler, you can use Result<T, E> as TryFromConn where T: TryFromConn<Error = E>

If extraction is infallible, implement FromConn.

Required Associated Types

type Error: Send + Sync + Sized + 'static

The Error type. Tf this is a Handler, you can extract Self directly in a ApiHandler signature, and Error will be called on Conn if try_from_conn fails.

Required Methods

fn try_from_conn<'life0, 'async_trait>( conn: &'life0 mut Conn ) -> Pin<Box<dyn Future<Output = Result<Self, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Attempt to extract Self from &mut Conn, returning Error in case of failure

Object Safety

This trait is not object safe.