Trait trillium_api::FromConn

pub trait FromConn: Send + Sync + Sized + 'static {
    // Required method
    fn from_conn<'life0, 'async_trait>(
        conn: &'life0 mut Conn
    ) -> Pin<Box<dyn Future<Output = Option<Self>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A trait to extract content from [Conn]s to be used as the second argument to an api handler. Implement this for your types.

Required Methods

fn from_conn<'life0, 'async_trait>( conn: &'life0 mut Conn ) -> Pin<Box<dyn Future<Output = Option<Self>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

returning None from this will not call the api handler, but will halt the conn.

Object Safety

This trait is not object safe.